parser-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pymupdf4llm==0.0.17
pymupdf==1.25.4
httpx==0.28.1
pydantic-settings==2.8.1
